Move function to styles util
